+++ CHANGES.md
@@ -99,6 +99,7 @@ Changes in 9.0
 - JS Hide Options textarea if Field not of select type in Fields.fnc.php
 - Add Balance widget in StudentBalances.php
 - Add Total sum of balances in StudentBalances.php
+- Fix SQL error check requested UserSyear & UserSchool exists in DB in Side.php, Search.fnc.php & SaveEnrollment.fnc.php
 
 Changes in 8.9.6
 ----------------